feat(openai-agents): add GenAI attribute constants (attributes-only branch)
