--- instrumentation-genai/opentelemetry-instrumentation-openai-agents/CHANGELOG.md
@@ -1,17 +1,0 @@
-# Changelog
-
-## Unreleased
-
-### Added
-
-- Initial implementation of OpenAI Agents instrumentation ([#3762](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3762))
-
-### Changed
-
-### Deprecated
-
-### Removed
-
-### Fixed
-
-### Security
