--- instrumentation-genai/opentelemetry-instrumentation-openai-agents/examples/.env.example
@@ -1,31 +1,0 @@
-# OpenAI Agents Instrumentation Configuration Example
-# ===================================================
-
-# Copy this file to your project and modify as needed
-# Set these environment variables to configure the instrumentation
-
-# Enable/disable content capture (request/response bodies)
-# Default: false
-
-# Standard OpenTelemetry configuration
-# Set your service name
-OTEL_SERVICE_NAME=my-openai-agents-service
-
-# Set your service version
-OTEL_SERVICE_VERSION=1.0.0
-
-# Set your resource attributes
-OTEL_RESOURCE_ATTRIBUTES=service.name=my-openai-agents-service,service.version=1.0.0
-
-# Configure exporters (example for OTLP)
-# OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4317
-# OTEL_EXPORTER_OTLP_HEADERS=api-key=your-api-key
-
-# Configure trace exporters
-# OTEL_TRACES_EXPORTER=otlp
-
-# Configure metrics exporters
-# OTEL_METRICS_EXPORTER=otlp
-
-# Configure logs exporters
-# OTEL_LOGS_EXPORTER=otlp
