--- instrumentation-genai/opentelemetry-instrumentation-openai-agents/examples/openai_agents_multi_agent_travel/.env.example
@@ -1,65 +1,0 @@
-# =============================
-# OpenAI / Azure OpenAI Config
-# =============================
-# Choose EITHER Azure OpenAI OR public OpenAI credentials.
-
-# --- Azure OpenAI (fill these if using Azure) ---
-AZURE_OPENAI_ENDPOINT=
-AZURE_OPENAI_API_KEY=
-AZURE_OPENAI_DEPLOYMENT=gpt-4.1-mini
-AZURE_OPENAI_API_VERSION=2024-08-01-preview
-
-# --- Public OpenAI (optional alternative) ---
-OPENAI_API_KEY=
-# Override model if desired (fallback in code is gpt-4o-mini)
-OPENAI_MODEL=
-
-# =============================
-# Application Insights (optional)
-# =============================
-# If provided, traces go to Azure Monitor via the connection string.
-APPLICATIONINSIGHTS_CONNECTION_STRING=
-# Legacy env name also supported
-APPLICATION_INSIGHTS_CONNECTION_STRING=
-
-# =============================
-# OpenTelemetry OTLP Export (optional)
-# =============================
-# Provide these to send data to another OTLP backend instead of (or in addition to) console.
-# OTEL_EXPORTER_OTLP_ENDPOINT=https://example-otlp-endpoint
-# OTEL_EXPORTER_OTLP_HEADERS=api-key=YOUR_KEY
-
-# =============================
-# Instrumentation Toggles
-# =============================
-# Content & tool capture may increase data volume. Set to 'false' to disable.
-OTEL_INSTRUMENTATION_OPENAI_AGENTS_CAPTURE_CONTENT=true
-OTEL_INSTRUMENTATION_OPENAI_AGENTS_CAPTURE_TOOL_IO=true
-OTEL_INSTRUMENTATION_OPENAI_AGENTS_CAPTURE_TOOL_DEFINITIONS=false
-OTEL_INSTRUMENTATION_OPENAI_AGENTS_CAPTURE_METRICS=true
-
-# =============================
-# Enhanced Planner Behavior
-# =============================
-# When 1 (default), enhanced.py prompts interactively if the coordinator
-# model returns clarifying questions instead of structured TripDetails.
-# Set to 0 to force a hard error (useful for CI / non-interactive runs).
-INTERACTIVE_TRIP_DETAILS=1
-
-# Application Insights Query (Logs API via AAD)
-APPLICATION_INSIGHTS_RESOURCE_ID=
-
-# Azure AD (if using LogsQueryClient)
-AZURE_TENANT_ID=
-AZURE_CLIENT_ID=
-AZURE_CLIENT_SECRET=
-
-# Evaluation script toggles
-EVAL_EMIT_OPERATION_DETAILS=1
-EVAL_EMIT_EVALUATIONS=1
-EVAL_VERIFICATION=1
-EVAL_THRESHOLD=3.0
-EVAL_LOOKBACK_MINS=30
-EVAL_MODEL_DEPLOYMENT=
-EVAL_MODEL=
-EVAL_MAX_MESSAGE_CHARS=20000
