--- instrumentation-genai/opentelemetry-instrumentation-openai-agents/examples/openai_agents_multi_agent_travel/requirements.txt
@@ -1,24 +1,0 @@
-python-dotenv
-openai>=1.40.0
-openai-agents>=0.3.2
-pydantic>=2.7.0
-
-# OpenTelemetry core & exporters
-opentelemetry-api
-opentelemetry-sdk
-opentelemetry-exporter-otlp
-azure-monitor-opentelemetry-exporter
-
-# Instrumentation for Agents API (install if published separately)
-opentelemetry-instrumentation-openai-agents>=0.1.0
-
-# For evaluation script
-azure-monitor-query          # If using LogsQueryClient mode
-azure-identity               # AAD auth for LogsQueryClient
-requests                     # REST fallback querying
-azure-ai-evaluation          # (optional) rich evaluators
-azure-monitor-opentelemetry-exporter
-opentelemetry-api
-opentelemetry-sdk
-opentelemetry-exporter-otlp
-python-dotenv
