--- instrumentation-genai/opentelemetry-instrumentation-openai-agents/examples/quickstart.py
@@ -1,71 +1,0 @@
-# pylint: skip-file
-import os
-
-from openai import OpenAI
-
-# NOTE: OpenTelemetry Python Logs and Events APIs are in beta
-from opentelemetry import _events, _logs, metrics, trace
-from opentelemetry.exporter.otlp.proto.grpc._log_exporter import (
-    OTLPLogExporter,
-)
-from opentelemetry.exporter.otlp.proto.grpc.metric_exporter import (
-    OTLPMetricExporter,
-)
-from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import (
-    OTLPSpanExporter,
-)
-from opentelemetry.instrumentation.openai_agents import (
-    OpenAIAgentsInstrumentor,
-)
-from opentelemetry.sdk._events import EventLoggerProvider
-from opentelemetry.sdk._logs import LoggerProvider
-from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
-from opentelemetry.sdk.metrics import MeterProvider
-from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchSpanProcessor
-
-# configure tracing
-trace.set_tracer_provider(TracerProvider())
-trace.get_tracer_provider().add_span_processor(
-    BatchSpanProcessor(OTLPSpanExporter())
-)
-
-# configure logging and events
-_logs.set_logger_provider(LoggerProvider())
-_logs.get_logger_provider().add_log_record_processor(
-    BatchLogRecordProcessor(OTLPLogExporter())
-)
-_events.set_event_logger_provider(EventLoggerProvider())
-
-# configure metrics
-metrics.set_meter_provider(
-    MeterProvider(
-        metric_readers=[
-            PeriodicExportingMetricReader(
-                OTLPMetricExporter(),
-            ),
-        ]
-    )
-)
-
-# instrument OpenAI with new instrumentor
-OpenAIAgentsInstrumentor().instrument()
-
-
-def main():
-    client = OpenAI()
-    chat_completion = client.chat.completions.create(
-        model=os.getenv("CHAT_MODEL", "gpt-4o-mini"),
-        messages=[
-            {
-                "role": "user",
-                "content": "Write a short poem on OpenTelemetry.",
-            },
-        ],
-    )
-    print(chat_completion.choices[0].message.content)
-
-
-if __name__ == "__main__":
-    main()
