+++ instrumentation-genai/opentelemetry-instrumentation-openai-agents/examples/weekend_planner_appinsights.py
@@ -1,0 +1,280 @@
+"""Weekend planner with OpenAI Agents + OpenTelemetry traces to Azure Application Insights.
+
+Prereqs:
+- pip install openai openai-agents opentelemetry-sdk azure-monitor-opentelemetry-exporter python-dotenv rich
+- Set APPLICATION_INSIGHTS_CONNECTION_STRING in your environment
+- For GitHub Models: set GITHUB_TOKEN (and optionally GITHUB_OPENAI_BASE_URL, GITHUB_MODEL)
+- For Azure OpenAI: set AZURE_OPENAI_ENDPOINT, AZURE_OPENAI_VERSION, AZURE_OPENAI_CHAT_DEPLOYMENT
+"""
+
+from __future__ import annotations
+
+import asyncio
+import logging
+import os
+import random
+from dataclasses import dataclass
+from datetime import datetime
+from typing import Callable
+from urllib.parse import urlparse
+
+import azure.identity
+import openai
+from agents import (
+    Agent,
+    OpenAIChatCompletionsModel,
+    Runner,
+    function_tool,
+    set_tracing_disabled,
+)
+from dotenv import load_dotenv
+from rich.logging import RichHandler
+
+from opentelemetry import trace
+from opentelemetry.instrumentation.openai_agents import (
+    OpenAIAgentsInstrumentor,
+)
+from opentelemetry.sdk.resources import Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import (
+    BatchSpanProcessor,
+    ConsoleSpanExporter,
+)
+
+load_dotenv(override=True)
+
+logging.basicConfig(
+    level=logging.WARNING,
+    format="%(message)s",
+    datefmt="[%X]",
+    handlers=[RichHandler()],
+)
+logger = logging.getLogger("weekend_planner")
+
+
+@dataclass
+class _ApiConfig:
+    """Helper describing how to create the OpenAI client."""
+
+    build_client: Callable[[], openai.AsyncOpenAI]
+    model_name: str
+    base_url: str
+    provider: str
+
+
+def _set_capture_env(provider: str, base_url: str) -> None:
+    """Set GenAI capture context (optional; capture defaults to enabled)."""
+    os.environ.setdefault("OTEL_GENAI_AGENT_NAME", "Weekend Planner Agent")
+    os.environ.setdefault(
+        "OTEL_GENAI_AGENT_DESCRIPTION",
+        "Assistant that plans weekend activities using weather and events data",
+    )
+    os.environ.setdefault("OTEL_GENAI_AGENT_ID", "weekend-planner")
+
+    parsed = urlparse(base_url)
+    if parsed.hostname:
+        os.environ.setdefault("OTEL_GENAI_SERVER_ADDRESS", parsed.hostname)
+    if parsed.port:
+        os.environ.setdefault("OTEL_GENAI_SERVER_PORT", str(parsed.port))
+
+
+def _resolve_api_config() -> _ApiConfig:
+    """Return the client configuration for the requested host."""
+
+    host = os.getenv("API_HOST", "github").lower()
+
+    if host == "github":
+        base_url = os.getenv(
+            "GITHUB_OPENAI_BASE_URL",
+            "https://models.inference.ai.azure.com",
+        ).rstrip("/")
+        model_name = os.getenv("GITHUB_MODEL", "gpt-4o")
+        api_key = os.environ["GITHUB_TOKEN"]
+
+        def _build_client() -> openai.AsyncOpenAI:
+            return openai.AsyncOpenAI(base_url=base_url, api_key=api_key)
+
+        return _ApiConfig(
+            build_client=_build_client,
+            model_name=model_name,
+            base_url=base_url,
+            provider="azure.ai.inference",
+        )
+
+    if host == "azure":
+        endpoint = os.environ["AZURE_OPENAI_ENDPOINT"].rstrip("/")
+        api_version = os.environ["AZURE_OPENAI_VERSION"]
+        deployment = os.environ["AZURE_OPENAI_CHAT_DEPLOYMENT"]
+
+        credential = azure.identity.DefaultAzureCredential()
+        token_provider = azure.identity.get_bearer_token_provider(
+            credential,
+            "https://cognitiveservices.azure.com/.default",
+        )
+
+        def _build_client() -> openai.AsyncAzureOpenAI:
+            return openai.AsyncAzureOpenAI(
+                api_version=api_version,
+                azure_endpoint=endpoint,
+                azure_ad_token_provider=token_provider,
+            )
+
+        return _ApiConfig(
+            build_client=_build_client,
+            model_name=deployment,
+            base_url=endpoint,
+            provider="azure.ai.openai",
+        )
+
+    raise ValueError(f"Unsupported API_HOST '{host}'")
+
+
+def _configure_otel() -> None:
+    """Configure tracer provider to export to both Console and Azure Monitor.
+
+    - Always adds ConsoleSpanExporter for local visibility.
+    - Adds AzureMonitorTraceExporter when APPLICATION_INSIGHTS_CONNECTION_STRING is set
+      and the azure-monitor-opentelemetry-exporter package is installed.
+    """
+
+    conn = os.getenv("APPLICATION_INSIGHTS_CONNECTION_STRING")
+    resource = Resource.create(
+        {
+            "service.name": "weekend-planner-service",
+            "service.namespace": "leisure-orchestration",
+            "service.version": os.getenv("SERVICE_VERSION", "1.0.0"),
+        }
+    )
+
+    tracer_provider = TracerProvider(resource=resource)
+
+    # Always enable console exporter
+    tracer_provider.add_span_processor(
+        BatchSpanProcessor(ConsoleSpanExporter())
+    )
+    print("[otel] Console span exporter configured")
+
+    # Optionally add Azure Monitor exporter
+    if conn:
+        try:
+            from azure.monitor.opentelemetry.exporter import (
+                AzureMonitorTraceExporter,
+            )
+        except ImportError:
+            print(
+                "[otel] Azure Monitor exporter not installed. "
+                "Install with: pip install azure-monitor-opentelemetry-exporter",
+            )
+        else:
+            tracer_provider.add_span_processor(
+                BatchSpanProcessor(
+                    AzureMonitorTraceExporter.from_connection_string(conn)
+                )
+            )
+            print("[otel] Azure Monitor trace exporter configured")
+    else:
+        print(
+            "[otel] APPLICATION_INSIGHTS_CONNECTION_STRING not set; skipping Azure Monitor"
+        )
+
+    trace.set_tracer_provider(tracer_provider)
+
+
+api_config = _resolve_api_config()
+_set_capture_env(api_config.provider, api_config.base_url)
+_configure_otel()
+
+OpenAIAgentsInstrumentor().instrument(
+    tracer_provider=trace.get_tracer_provider()
+)
+
+client = api_config.build_client()
+set_tracing_disabled(False)
+
+
+@function_tool
+def get_weather(city: str) -> dict[str, object]:
+    logger.info("Getting weather for %s", city)
+    if random.random() < 0.05:
+        return {
+            "city": city,
+            "temperature": 72,
+            "description": "Sunny",
+        }
+    return {
+        "city": city,
+        "temperature": 60,
+        "description": "Rainy",
+    }
+
+
+@function_tool
+def get_activities(city: str, date: str) -> list[dict[str, object]]:
+    logger.info("Getting activities for %s on %s", city, date)
+    return [
+        {"name": "Hiking", "location": city},
+        {"name": "Beach", "location": city},
+        {"name": "Museum", "location": city},
+    ]
+
+
+@function_tool
+def get_current_date() -> str:
+    """Gets the current date and returns as YYYY-MM-DD."""
+    logger.info("Getting current date")
+    return datetime.now().strftime("%Y-%m-%d")
+
+
+agent = Agent(
+    name="Weekend Planner",
+    instructions=(
+        "You help users plan their weekends and choose the best activities for the given weather. "
+        "If an activity would be unpleasant in the weather, don't suggest it. Include the date of the weekend in your response."
+    ),
+    tools=[get_weather, get_activities, get_current_date],
+    model=OpenAIChatCompletionsModel(
+        model=api_config.model_name, openai_client=client
+    ),
+)
+
+
+def _root_span_name(provider: str) -> str:
+    return f"weekend_planning_session[{provider}]"
+
+
+async def main() -> None:
+    tracer = trace.get_tracer(__name__)
+    with tracer.start_as_current_span(
+        _root_span_name(api_config.provider)
+    ) as span:
+        user_request = "What can I do this weekend in Seattle?"
+
+        span.set_attribute("user.request", user_request)
+        span.set_attribute("api.host", os.getenv("API_HOST", "github"))
+        span.set_attribute("model.name", api_config.model_name)
+        span.set_attribute("agent.name", agent.name)
+        span.set_attribute("target.city", "Seattle")
+
+        try:
+            result = await Runner.run(agent, input=user_request)
+            print(result.final_output)
+
+            span.set_attribute(
+                "agent.response",
+                result.final_output[:500] if result.final_output else "",
+            )
+            span.set_attribute("request.success", True)
+            logger.info("Weekend planning completed successfully")
+        except Exception as exc:
+            span.record_exception(exc)
+            span.set_attribute("request.success", False)
+            logger.error("Error during weekend planning: %s", exc)
+            raise
+
+
+if __name__ == "__main__":
+    logger.setLevel(logging.INFO)
+    try:
+        asyncio.run(main())
+    finally:
+        trace.get_tracer_provider().shutdown()
